extract interface loaders logic to helper
