+++ apps/web-remix/app/components/pages/pipelines/interface/websiteChatbot/loader.server.tsx
@@ -1,32 +1,13 @@
-import { json } from '@remix-run/node';
 import type { LoaderFunctionArgs } from '@remix-run/node';
-import invariant from 'tiny-invariant';
 
-import { PipelineApi } from '~/api/pipeline/PipelineApi';
+import { interfaceLoader } from '~/components/pages/pipelines/interface/interface.loader.server';
 import { requireLogin } from '~/session.server';
 import { loaderBuilder } from '~/utils.server';
 
 export async function loader(args: LoaderFunctionArgs) {
-  return loaderBuilder(async ({ request, params }, { fetch }) => {
+  return loaderBuilder(async ({ request, params, ...rest }, helpers) => {
     await requireLogin(request);
-    invariant(params.organizationId, 'organizationId not found');
-    invariant(params.pipelineId, 'pipelineId not found');
 
-    const pipelineApi = new PipelineApi(fetch);
-    const aliasId = pipelineApi.getAliasFromUrl(request.url);
-
-    const pipeline = await pipelineApi.getAliasedPipeline(
-      params.organizationId,
-      params.pipelineId,
-      aliasId,
-    );
-
-    return json({
-      pipeline,
-      aliasId,
-      pipelineId: params.pipelineId,
-      organizationId: params.organizationId,
-      pageUrl: process.env['PAGE_URL'],
-    });
+    return interfaceLoader({ request, params, ...rest }, helpers);
   })(args);
 }
